Started ServerApplication
ServerApplication
- Reads data from json file
- Returns data as json to the client

DataCollector and ServerApplication have configs that they can read
from a file (config.json) and generate the file in case it doesn't exist
or it's invalid. DC and SA both have a separate config.py that does
exactly the same thing. This should be fixed so that both apps can have
their own configurable settings but still use a common module.