Improved blender script + svg to json converter
Svg to json converter is very simple and doesn't work for every svg.
It expects every tag in the svg to be on their own line.

The blender script takes a json file that is outputted by the
svg to json converter and creates meshes from it.
It's also very simple but works (somewhat).